Observe cros-disks FormatCompleted signal instead of FormattingFinished.
The FormattingFinished signal has been long deprecated (back in 2011)
and replaced by the FormatCompleted signal, which allows cros-disks to
properly notify Chrome the completion of a format operation with an
appropriate error code. This CL modifies CrosDisksClient and
DiskMountManager to use the FormatCompleted signal.

BUG=369877
TEST=Tested the following:
1. Run chromeos_unittests.
2. Insert a USB drive and format the mounted partition via Files.app.
   Verify that Files.app notifies the successful completion of the
   format operation.
3. Insert a write-protected SD card and format the mounted partition via
   Files.app. Verify that Files.app reports an error.

R=stevenjb@chromium.org, tbarzic@chromium.org

Review URL: https://codereview.chromium.org/266143002