WebRTC: Improve virtual webcam check.
The script currently reports a success incorrectly
when the v4l2loopback module is not properly loaded on Linux.
This CL at least catches such an error, even if it's
an ugly solution it's still an improvement.
We plan to move away from having virtual webcams but I
still wanted to fix this for now.

TEST=local testing running the script.
BUG=

Review URL: https://codereview.chromium.org/260683002