Add a regression test for closing the app list natively.
Usually the app list is just dismissed and the window kept around. On
Windows and Linux, a close request can come from the OS by
right-clicking the taskbar/shelf. There is no way to prevent the window
closing in this case, so it is torn down. This exercises a bunch of
codepaths that don't currently have tests.

BUG=370385,169114

Review URL: https://codereview.chromium.org/266423004