Fix cleaning up of cancelled reparents in the Windows App List Folder UI.
This CL fixes the drag cancellation behavior for the app list on Windows.
Previously, the drag view in the root level grid view would not be cleaned
up causing an invalid AppListItemView to persist in the root level view.
When interacted with, this would crash.

BUG=365600

Review URL: https://codereview.chromium.org/262093004