NaCl non-sfi sandbox: use modern framework for tests
The non-sfi sandbox policy is the implementation of the SandboxBPFPolicy
interface but tests are testing it as an old-style function pointer
policy.
This CL makes use of BPF_TEST_C and BPF_DEATH_TEST_C and takes a classname
as a parameter rather than a function pointer.

BUG=359285
R=mseaborn@chromium.org

Review URL: https://codereview.chromium.org/269333004