Solid Color Scrollbars should start with opacity 0.
Recent changes to scrollbar animation have caused pinch-zoom overlay scrollbars
to once again be visible when page_scale_factor == 1. The mechanism landed in
r266007 seems to work except for the inital value of the opacity of the
solid color scrollbar. This CL forces SolidColorScrollbarLayerImpl layers to
be created with opacity 0.

BUG=369688

Review URL: https://codereview.chromium.org/265043005