[Files.app] Cache urls of entries before sorting.
toURL() consumes about a half time of filelist draw. In sort, toURL is called by the compare function, which is called many times in sort.

The great solution is making toURL faster, but as for now, we cache urls and use it in sort, instead of calling toURL in each call.

BUG=367123
TEST=manually.
R=hirono@chromium.org

Review URL: https://codereview.chromium.org/270573004