Password bubble: Deal correctly with blacklist state changes.
If a user blacklists or unblacklists a site in a separate tab, we need
to correctly update every other tab's state (consider the unlikely case
of logging into the same site 8 times at the same time in separate tabs,
or the slightly less unlikely case of changing status in
chrome://settings/passwords while a login page is visible). This CL updates
the observer logic to account for blacklist entries being added and removed
from the PasswordStore, and updates the UI accordingly.

BUG=261628
R=vabr@chromium.org

Review URL: https://codereview.chromium.org/276683003