Add support for more ASAN errors generation to chrome://crash/...
This is a clone of https://codereview.chromium.org/247543002/ that don't crash in debug.

Adds the following crash url to chrome/chromium builds when SYZYASAN is defined (built with syzyasan=1 in GYP_DEFINES):
heap-corrupted-block

This allows easy validation of the instrumentation/reporting for a given build without having to track down a known/unfixed memory error.

R=chrisha@chromium.org, jochen@chromium.org

BUG=

Review URL: https://codereview.chromium.org/267773005