First step towards invalidations componentization
Moves some files from chrome/browser/invalidation to
components/invalidation.  The files moved in this CL are those that do
not have dependencies on chrome/, so they can be moved without
significant refactoring.

This refactoring touches a lot of files, but none of the changes are
meaningful.  Most of it inovlves changing #include paths, updating DEPS,
changing header guards, and other trivial changes.

One part of this change that is significant is that it creates two new
static libraries and moves some of the symbols formerly defined in
chrome/browser/invalidation into those libraries.  There is some risk of
linker errors if this was done wrong.

TBR=bartfab,blundell
BUG=336571

Review URL: https://codereview.chromium.org/255273002