[Mac] Use full icon for small Chrome Apps custom folder assets
-[NSWorkspace setIcon:] has a bug (rdar://16129816) handling
premultiplication when using named NSImages. This is most noticeable for
the small assets, since there is a higher proportion of pixels with an
alpha component.

We still want to use a named image for the larger icons because the
folder asset is textured and doesn't compress well. This CL changes the
16x16 and 32x32 assets to use a pre-rendered icon, rather than an
overlay.

BUG=305373

Review URL: https://codereview.chromium.org/270623003