Linux Aura Task Manager Frame Buttons Misaligned
The task manager is a CustomFrameView. This had the logic for its button
placement hardcoded. The view now adds itself as a listener for button
configuration changes on linux. For other systems it falls back on a
default set that matches the current (minimize, maximize, close) order.

The layout of the buttons has changed to consult this order definition.

Loading the assets for the close button has been moved out of layout.
ResourceBundle only loads the asset once, so this can be done during
initialization.

The top padding for the buttons has been updated to provide more spacing. It now fits the rest of the padding in the class.

TEST=CustomFrameViewTest
BUG=351917

Review URL: https://codereview.chromium.org/240163006