Enable auto-refreshing of the gcm-internals page whenever a GCM activity is recorded, and fix a bug that previously causes this cl to crash the browser upon signin.
Basically GCMStatsRecorder::RecordRegistrationResponse() method did not check whether is_recording_ is true or not, so it always records. When the user signs in, a registration event is automatically triggered, but since the gcm-internals page has not been loaded yet, a callback object is still NULL. Therefore recorder calling this callback will crash the browser. I have added the check as well as test case for this scenario.

BUG=341256, 371259

Review URL: https://codereview.chromium.org/277743002