ozone: Make the evdev implementation a PlatformEventSource.
Instead of using the EventFactoryOzone as the PlatformEventSource, allow
ozone's platform implementation to decide what to use as the event-source.
For the evdev ozone-platform, the event-factory (EventFactoryEvdev) remains
as the event source. In the case of X11 ozone-platform, it will create the
X11EventSource (currently used on chromeos and linux-aura) as the
event-source.

BUG=361137
R=spang@chromium.org

Review URL: https://codereview.chromium.org/280033002