views: Terminate the nested message-loop correctly when closing the menu.
It is possible for a non-menu source to start a nested message loop
(e.g. for dragging a menu-item). So attempting to terminate the current
iteration of the message-loop when the menu is closed can be problematic.
So instead, terminate the message-loop through the DispatcherClient.

BUG=367850, 372208
R=sky@chromium.org

Review URL: https://codereview.chromium.org/279073002