Switch thread_local to use ThreadPlatformStorage in Android
It might not be good to switch thread_local to use ThreadPlatformStorage
on all platforms, in order to pass all tests in Linux, the number of TLS
slot has be 1024, it means 4k memory will be temporarily increased in stack,
and then move to heap for every thread which use the TLS.

Android only needs 256 slot which is much smaller than Linux's usage, and
TLS slot is currently only running out on Android's tests.

So this patch is only for Android platform.

BUG=264406

Review URL: https://codereview.chromium.org/189263003