Allow option to ignore default route
In OpenVPN, it is up to the client to decide whether or not to
create a default route through the OpenVPN server.  Even when the
server supplies a gateway address, it isn't always clear if the
intent is for the client to set a default route through this
gateway.

By default, ChromeOS creates a default route to the gateway.
This CL adds an ONC configuration option to allow ONC to override
this default behavior, thus providing the means for a "split
tunnel" configuration in situations where the route configuration
provided by the VPN server is amenable to this.  Supplying this
option does not override any explicit routes pushed by the server.

BUG=370460
R=pneubeck@chromium.org

Review URL: https://codereview.chromium.org/275723002