Makes BackgroundServiceLoader destroy loader on background thread
As the loader is used on the background thread it should also be
destroyed on the background thread. This way all usage is kept to a
single thread so that if the loader uses resources that require access
on the same thread we don't run into issues.

BUG=372585
TEST=covered by tests
R=davemoore@chromium.org

Review URL: https://codereview.chromium.org/276573013