Android: Move scheduling logic to CompositorImpl
This adds two ways of scheduling:

- implicitly from LTHClient::ScheduleComposite():
This will post a task to composite immediately.

- explicitly (for Chrome) through SetNeedsComposite() where
there are no layer tree changes, but instead a Composite()
will be scheduled for vsync with the Layout() callback giving
the app a chance to update the tree.

TBR=skyostil@chromium.org
NOTRY=True

BUG=234173

Review URL: https://codereview.chromium.org/239963002