Add support for building Ubuntu/Trusty sysroot images.
The main reason for this is to support cross building for
ARM on Trusty machines (where the sysroot should match the
cross compiler version).

Rename the main sysroot creator script and source it in the
trust and wheezy scripts that are actually runnable.

This should make it easy to create a script for Debian/Jesse
when it is released without duplicating code.

Review URL: https://codereview.chromium.org/271853003