runtest.py: Add flags for memory tools.
Add flags for a variety of functions that was previously
controlled with factory properties.
This will make it easier to use these tools from recipes.

NOTICE: This CL fixes the --enable-lsan flag so it no longer
can be used as --enable-lsan=True (which was previously
required). From now on only --enable-lsan works.

BUG=None
TEST=Local execution with various flags.

Review URL: https://codereview.chromium.org/278743004