Add get_syzygy_binaries.py script.
This adds a script for retrieving and installing Syzygy binaries. Currently binaries are checked in to an SVN repository and installed via a DEPS dependency. The Syzygy team is moving to git, and archiving their binaries on Google Common Data Storage. A later CL will move to using this script, installing the binaries via a DEPS hook.

BUG=
R=siggi@chromium.org

Review URL: https://codereview.chromium.org/288703002