Tighten expectations in RenderViewImplTest::SendProgressCompletionEvent.
This test is failing on the TSANv2 bot. From analysis it seems that an
assumption about the number of messages sent might not match reality
under some circumstances. To test that theory this change encodes
that assumption into the test. If the theory is true the new expectation
should fail on the failing bot.

TBR=avi
BUG=371865

Review URL: https://codereview.chromium.org/287513004