Randomize sync point starting number to reduce risk of GPU hangs.
There's a race condition in sync point creation where code can start a wait on a sync point immediately before the sync point is inserted. This will cause the GPU channel to lock up, as the retire will be placed in the queue after the wait.
This normally won't happen in practice (as the client should only wait on sync points that have been inserted) but it can happen if the GPU process restarts and the client waits on a sync point created by the old GPU process.

To reduce the risk of this happening, randomize the starting sync point ID. That way it's much less likely that the sync point created by the old GPU process will soon be created in the new GPU process.

BUG=365904,373452

Review URL: https://codereview.chromium.org/279303003