sync: Improve handling of bad UniquePos (retry)
Retry: The first attempt had a memory leak.  It seems that this leak existed
before this CL, but no code exercised it.  This updated CL includes a scoped
deleter that should fix it.

Original commits message was:

Makes the client assign a valid position to incoming bookmarks if the
server has not populated the required fields.  This code should never be
triggered unless there is a bug in the server.  This risks reordering
users' bookmarks, but that's probably preferable to a crash.  This
fallback code is still protected by a NOTREACHED(), as before.

Detects bookmarks that do not have valid position information during
database load.  If these corrupted bookmarks are detected, the entire
database is declared to be corrupt.  Sync will then re-download
all of the user's data, which should fix the problem.

BUG=367247

Review URL: https://codereview.chromium.org/283143002