DevTools test harness: add logToStderr method on testRunner.
it is very hard to troubleshoot a flaky asynchronous DevTools test without
being able to log something directly to the console output.
With this patch DevTools window is able to do that. Also logging into
Error stream doesn't skew the normal test output so the flaky test
could be tested in a cycle until the first error.

BUG=

Review URL: https://codereview.chromium.org/288663003