ozone: Remove the explicit call to start listening for events.
It is not necessary for WindowTreeHostOzone to call into the event-factory
to explicitly request it to start receiving events. The evdev implementation
can start listening to the events immediately, and the canca implementation
can start listening for events when the dispatcher list changes (this is the
same we do for the libevent based x11 event-source).

BUG=361137
R=spang@chromium.org

Review URL: https://codereview.chromium.org/285303004