Store the service worker script and its imports on first load... kinda
The resource data is put into an in-memory instance of a net::DiskCache for now. That cache will be used for (some) subsequent starts of the worker in the current browsing session, but will not be used across browser restarts. The next step is to have the registration job store the resource ids with the registration data on disk and to also put the resource data to disk.

BUG=364318

Review URL: https://codereview.chromium.org/269373002