PPAPI: Make generator support "no_interface_string"
This is so I can introduce a new way to do PPP interfaces that are
passed to a related PPB interface, instead of using PPP_GetInterface.

Note this doesn't add support for the PNaCl shim yet, although it should
at least make it compile.

BUG=367896

Review URL: https://codereview.chromium.org/281203002