VS2013 Update 2
As discussed in the linked bug, the toolchain2013.py script in
depot_tools is no longer workable (because the update for Update 2 is
supplied as a .msp that must be applied against a system-installed
VS2013).

As such, the Express hash here is not updated.

The hash zip referenced here was built as follows:
- Install VS2013 Update 2 on a clean VM
- Copy DIA SDK\, VC\
- Copy DLLs from VC\redist to sys32\ and sys64\
- Copy win8sdk unchanged from previous .zip.
- Delete various unused arm\ subdirectories
- Delete the IDE-only Snippets, etc. subdirectories in VC\
- Patch VC\include\xtree to disable warning 4702 per request in
  http://crbug.com/346399 .

A followup change will be to write a script that does these steps
semi-automatically, hopefully for Express too, though there is the
added complication of the WDK/ATL/MFC hacking required there. In
particular, this script will not be useful for a dev to actually run
as part of runhooks, but will be a bit useful for deployment/
documentation of the above process.

In the interim there are no extremely-pressing reasons for Express
users to update to Update2 that I'm aware of, so they get non-Update2
for now.

R=iannucci@chromium.org
BUG=372451,346399,371847,339215,350639

Review URL: https://codereview.chromium.org/284663003