Fix virtual keyboard.
Virtual keyboard starts with a height of 0, which makes us wait for a frame at
that size which will never come.
RWHI will skip the resizes for this, hence will not throttle the next (correct)
size, which will come in but then be discarded.

Make sure not to create a resize lock for empty sizes.

BUG=374196

Review URL: https://codereview.chromium.org/284413002