Switch to using tab_capture_end2end_tests isolate in GPU recipe.
This isolate doesn't use xvfb on Linux, which was causing problems
with the browser_test isolate. Re-enabled this test on Linux.

It turns out to be more work than it's worth to make this test run both via
the isolate and as a standalone executable. The plan is to switch the GPU bots
to use isolates everywhere imminently, so for the moment, run the new test
only via the isolate.

Tested locally on Linux. Isolate runs successfully through the recipe.

BUG=365927

Review URL: https://codereview.chromium.org/286933016