Remove override of View::HitTestRect() from ImageView
The override of View::HitTestRect() in ImageView forces hit
testing to fail if the |interactive_| member is false, which
prevents the view from accepting events. Since the |interactive_|
member is unrelated to location, it is better to move this logic
into an override of EventTarget::CanAcceptEvent() instead.

BUG=374325
TEST=none

Review URL: https://codereview.chromium.org/282403002