Moved IS_EPHEMERAL flag to extension prefs
This patch moves the IS_EPHEMERAL flag from the Extension class to
extension prefs. This is because an Extension object is immutable and
this flag needs to be changed after creation to quickly promote a
running ephemeral app to being installed. Otherwise the install must
be delayed until the ephemeral app is closed.

BUG=374018
TBR=sky@chromium.org
TEST=All automated tests

Review URL: https://codereview.chromium.org/282103003