Fixes for getting themes to work well with Windows High DPI.
Relanding this as the previous attempt was reverted due to compile failures due to the IsScaleFactorSupported
function being moved to resource_bundle.h/.cc. Fixed the ctor for BrowserThemePack to workaround it.

The changes are to use the GetScaleForScaleFactor function to get the scale instead
of GetImageScale. The latter function returns the current display scale on Windows which
causes themes to break. The other change is in the ctor of the BrowserThemePack class
where we add SCALE_FACTOR_100P to the list of supported scales as it is needed for default
theme packs to work correctly.

BUG=362979
R=oshima
TBR=pkotwicz

Review URL: https://codereview.chromium.org/291823006