Implement Android accessible hit testing using an IPC to the renderer process.
Only Blink can accurately determine what node is under the user's finger.
Implement this by sending a hit test request every time the user does
touch exploration, and returning a hover event on the appropriate accessible
node in response.

The test is meant to be a good sanity test and a skeleton, it tests that
the hit test request returns a hover event in response, and I intend to
follow this up with more specific tests of behavior as we tweak the
algorithm.

BUG=357371

Review URL: https://codereview.chromium.org/290633002